Disable all failing mojo specs and remove their schedules
This disables the jenkins jobs for the bulk of the mojo specs, and
it removes the timed trigger schedule for the disabled jobs.

If and when the individual specs are passing, they can be re-enabled
and re-scheduled at that time.  The logic behind this is that it makes
no sense to spin cloud compute resources for known-failing, and sizable
deployment tests.

The specs that remain enabled are:  magpie, ceph base, and designate ha.