Pin pip and virtualenv to <20.3 and <20.0
Also set the minimum version for tox to be 3.2.0.  This is to ensure
that a) we continue py27 compatibility, b) we don't introduce breaking
changes in pip (losing py35 on xenial) and c) the resolver continues to
work the way it used to whilst the technical debt is resolved in the
dependencies.