Disable charms for charmhub migration experiments
As part of the charmhub migration two charms have been selected to be
'split' so that charmhub publishing can be verified.  Thus, the pushers
in the Jenkins CI are being disabled so that the charmstore is no longer
updated for those two charms.  This is to avoid a 'split-brain' syndrome
of incorrect information between the two stores.